% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econagent-package.R
\docType{package}
\name{econagent-package}
\alias{econagent}
\alias{econagent-package}
\title{econagent: Utility Functions and Composite Goods in Economics}
\description{
Provides utility functions and composite goods for building economic models.
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
