% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econgoods-package.R
\docType{package}
\name{econgoods-package}
\alias{econgoods}
\alias{econgoods-package}
\title{econgoods: Utility Functions and Composite Goods in Economics}
\description{
Provides utility functions and composite goods for building economic models.
}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
