% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goods-consume.R
\name{goods_consume}
\alias{goods_consume}
\title{Consume goods}
\usage{
goods_consume(data, incomes = NULL)
}
\arguments{
\item{data}{A \code{econ_goods} object.}

\item{incomes}{A data frame of a list of data frames that contains columns
\code{income}.}
}
\value{
A \code{econ_goods} object.
}
\description{
Consume goods
}
