% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goods-consume.R
\name{goods_consume_recursively}
\alias{goods_consume_recursively}
\title{Consume goods recursively}
\usage{
goods_consume_recursively(data, f, ...)
}
\arguments{
\item{data}{A \code{econ_goods} object.}

\item{f}{A function that returns a data frame with columns \code{income}.}

\item{...}{Additional arguments passed to \code{stats::optim()}.}
}
\value{
A \code{econ_goods} object.
}
\description{
Consume goods recursively
}
