% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goods-produce.R
\name{goods_produce}
\alias{goods_produce}
\title{Produce goods}
\usage{
goods_produce(data, quantities = NULL, gradient = FALSE)
}
\arguments{
\item{data}{A \code{econ_goods} object.}

\item{quantities}{A data frame of a list of data frames that contains columns
\code{quantity}.}

\item{gradient}{Logical input to return the gradient. By default, \code{FALSE}.}
}
\value{
A \code{econ_goods} object.
}
\description{
Produce goods
}
