% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goods-produce.R
\name{goods_produce_recursively}
\alias{goods_produce_recursively}
\title{Produce goods recursively}
\usage{
goods_produce_recursively(data, f, ...)
}
\arguments{
\item{data}{A \code{econ_goods} object.}

\item{f}{A function that returns a data frame with columns \code{quantity}.}

\item{...}{Additional arguments passed to \code{stats::optim()}.}
}
\value{
A \code{econ_goods} object.
}
\description{
Produce goods recursively
}
