% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goods-reprice.R
\name{goods_reprice_recursive}
\alias{goods_reprice_recursive}
\title{Reprice goods recursively}
\usage{
goods_reprice_recursive(data, f, gradient = NULL, ...)
}
\arguments{
\item{data}{A \code{econ_goods} object.}

\item{f}{A function that returns a data frame with columns \code{price}.}

\item{gradient}{A function that returns a data frame with columns \code{prices}.}

\item{...}{Additional arguments passed to \code{stats::optim()}.}
}
\value{
A \code{econ_goods} object.
}
\description{
Reprice goods recursively
}
