% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goods-trade.R
\name{goods_trade_update}
\alias{goods_trade_update}
\title{Update trade cost}
\usage{
goods_trade_update(data, trade)
}
\arguments{
\item{data}{A \code{econ_goods} object.}

\item{trade}{A data frame with \code{cost} column.}
}
\value{
A \code{econ_goods} object.
}
\description{
Update trade cost
}
