% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goods-update.R
\name{goods_update}
\alias{goods_update}
\title{Update goods data}
\usage{
goods_update(data, value)
}
\arguments{
\item{data}{A \code{econ_goods} object.}

\item{value}{A data frame with new values.}
}
\value{
A \code{econ_goods} object.
}
\description{
Update goods data
}
