% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{new_util}
\alias{new_util}
\title{Create a new utility function}
\usage{
new_util(f, ..., class = character())
}
\arguments{
\item{f}{A function that returns the utility.}

\item{...}{Parameters to be passed to \code{f}.}

\item{class}{Name of subclass.}
}
\value{
A \code{econ_util} object.
}
\description{
Create a new utility function
}
