% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-homothetic.R
\name{new_util_homothetic}
\alias{new_util_homothetic}
\title{Create a new homothetic utility function}
\usage{
new_util_homothetic(f, ..., class = character())
}
\arguments{
\item{f}{A function that returns the utility and its gradient.
It has a \code{grad} argument that allows logical input.}

\item{...}{Parameters to be passed to \code{f}.}

\item{class}{Name of subclass.}
}
\value{
A \code{util_homothetic} object.
}
\description{
Create a new homothetic utility function
}
