% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-2goods.R
\name{util_2goods_budget}
\alias{util_2goods_budget}
\title{Budget line function factory for two goods}
\usage{
util_2goods_budget(prices, income)
}
\arguments{
\item{prices}{A numeric vector of length 2 with the prices of the goods.}

\item{income}{A scalar numeric of income.}
}
\value{
A function that takes a scalar numeric of quantity of good X and
returns a scalar numeric of quantity of good Y.
}
\description{
Budget line function factory for two goods
}
