% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-2goods.R
\name{util_2goods_indifference}
\alias{util_2goods_indifference}
\title{Indifference curve function factory for two goods}
\usage{
util_2goods_indifference(
  f,
  utility,
  otherwise = NA_real_,
  interval = c(1e-06, 1e+06),
  tol = 1e-06,
  ...
)
}
\arguments{
\item{f}{A \code{econ_util} object.}

\item{utility}{A scalar numeric of utility.}

\item{otherwise}{Default value when the root is not found. By default,
\code{NA_real}.}

\item{interval}{Passed to \code{\link[stats:uniroot]{stats::uniroot()}}.}

\item{tol}{Passed to \code{\link[stats:uniroot]{stats::uniroot()}}.}

\item{...}{Passed to \code{\link[stats:uniroot]{stats::uniroot()}}.}
}
\value{
A function that takes a scalar numeric of quantity of good X and
returns a scalar numeric of quantity of good Y.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
