% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-2goods.R
\name{util_2goods_utility}
\alias{util_2goods_utility}
\title{Utility function factory for two goods}
\usage{
util_2goods_utility(f, quantity_y, gradient = FALSE)
}
\arguments{
\item{f}{A \code{util} object.}

\item{quantity_y}{A scalar numeric of quantity of good Y.}

\item{gradient}{Logical input to return the gradient. By default, \code{FALSE}.}
}
\value{
A function that takes a scalar numeric of quantity of good X and
returns a scalar numeric of total utility (\code{gradient = TRUE}) or marginal
utility (\code{gradient = FALSE}).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
