% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-ces.R
\name{util_ces}
\alias{util_ces}
\alias{util_ces2}
\title{Constant elasticity of substitution (CES) utility function}
\usage{
util_ces(
  substitution,
  homogeneity = 1,
  efficiency = NA_real_,
  weights = double(),
  homothetic = homogeneity == 1
)

util_ces2(
  elasticity_of_substitution,
  homogeneity = 1,
  efficiency = NA_real_,
  weights = double(),
  homothetic = homogeneity == 1
)
}
\arguments{
\item{substitution}{A scalar numeric of substitution parameter.}

\item{homogeneity}{A scalar numeric of degree of homogeneity. By default,
\code{1}. When \code{homothetic = TRUE}, \code{homogeneity} must be equal to \code{1}.}

\item{efficiency}{A scalar numeric of efficiency parameter.}

\item{weights}{A numeric vector of weights.}

\item{homothetic}{A logical scalar. By default, \code{homogeneity == 1}.}

\item{elasticity_of_substitution}{A scalar numeric of elasticity of
substitution.}
}
\value{
A \code{util_ces} object.
}
\description{
Constant elasticity of substitution (CES) utility function
}
