% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-cet.R
\name{util_cet}
\alias{util_cet}
\alias{util_cet2}
\title{Constant elasticity of transformation (CET) function}
\usage{
util_cet(
  substitution,
  homogeneity = 1,
  efficiency = NA_real_,
  weights = double(),
  homothetic = homogeneity == 1
)

util_cet2(
  elasticity_of_transformation,
  homogeneity = 1,
  efficiency = NA_real_,
  weights = double(),
  homothetic = homogeneity == 1
)
}
\arguments{
\item{substitution}{A scalar numeric of substitution parameter. Typically
greater than \code{1}, so that the transformation surface is convex.}

\item{homogeneity}{A scalar numeric of degree of homogeneity. By default,
\code{1}. When \code{homothetic = TRUE}, \code{homogeneity} must be equal to \code{1}.}

\item{efficiency}{A scalar numeric of efficiency parameter.}

\item{weights}{A numeric vector of weights.}

\item{homothetic}{A logical scalar. By default, \code{homogeneity == 1}.}

\item{elasticity_of_transformation}{A scalar numeric of elasticity of
transformation.}
}
\value{
A \code{util_cet} object.
}
\description{
Constant elasticity of transformation (CET) function
}
