% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-cobb-douglas.R
\name{util_cobb_douglas}
\alias{util_cobb_douglas}
\title{Cobb-Douglas utility function}
\usage{
util_cobb_douglas(efficiency = NA_real_, weights = double(), homothetic = TRUE)
}
\arguments{
\item{efficiency}{A scalar numeric of efficiency parameter. By default,
\code{NA_real_}.}

\item{weights}{A numeric vector of weight parameters. By default, \code{double()}.}

\item{homothetic}{A logical scalar. By default, \code{TRUE}.}
}
\value{
A \code{util_cobb_douglas} object.
}
\description{
Cobb-Douglas utility function
}
