% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util_demand}
\alias{util_demand}
\title{Demand function}
\usage{
util_demand(f, prices, income = NULL, utility = NULL, grad = FALSE, ...)
}
\arguments{
\item{f}{A \code{util} object.}

\item{prices}{A numeric vector of prices.}

\item{income}{A scalar numeric of income. If \code{NULL}, \code{utility} must be
provided.}

\item{utility}{A scalar numeric of utility level. If \code{NULL}, \code{income} must be
provided.}

\item{grad}{Logical input to return the gradient. By default, \code{FALSE}.}

\item{...}{Additional arguments.}
}
\value{
when \code{grad = FALSE}, a numeric vector of quantities. When
\code{grad = TRUE}, a numeric matrix of gradients of quantities related to prices.
}
\description{
This function works as Marshallian demand function when \code{income} is input, and
as Hicksian demand function when \code{utility} is input.
}
