% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util_demand_marshallian}
\alias{util_demand_marshallian}
\title{Marshallian demand function}
\usage{
util_demand_marshallian(f, prices, income, gradient = FALSE, ...)
}
\arguments{
\item{f}{A \code{econ_util} object.}

\item{prices}{A numeric vector of prices.}

\item{income}{A scalar numeric of income.}

\item{gradient}{Logical input to return the gradient. By default, \code{FALSE}.}

\item{...}{Additional arguments.}
}
\value{
when \code{gradient = FALSE}, a numeric vector of quantities. When
\code{gradient = TRUE}, a numeric matrix of gradients of quantities related to prices.
}
\description{
Marshallian demand function
}
