% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util_gradient}
\alias{util_gradient}
\title{Gradient of a utility function}
\usage{
util_gradient(f, quantities, ...)
}
\arguments{
\item{f}{A \code{econ_util} object.}

\item{quantities}{A numeric vector of quantities.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector of gradients.
}
\description{
Gradient of a utility function
}
