% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-leontief.R
\name{util_leontief}
\alias{util_leontief}
\title{Leontief utility function}
\usage{
util_leontief(efficiency = NA_real_, weights = double())
}
\arguments{
\item{efficiency}{A scalar numeric of efficiency parameter. By default,
\code{NA_real_}.}

\item{weights}{A numeric vector of weight parameters. By default, \code{double()}.}
}
\value{
A \code{util_leontief} object.
}
\description{
Leontief utility function
}
