% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-linear.R
\name{util_linear}
\alias{util_linear}
\title{Linear utility function}
\usage{
util_linear(efficiency = NA_real_, weights = double())
}
\arguments{
\item{efficiency}{A scalar numeric of efficiency parameter. By default,
\code{NA_real_}.}

\item{weights}{A numeric vector of weight parameters. By default, \code{double()}.}
}
\value{
A \code{util_linear} object.
}
\description{
Linear utility function
}
