% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util_price}
\alias{util_price}
\title{Price function}
\usage{
util_price(f, prices, quantities, ...)
}
\arguments{
\item{f}{A \code{econ_util} object.}

\item{prices}{A numeric vector of prices.}

\item{quantities}{A numeric vector of quantities.}

\item{...}{Additional arguments.}
}
\value{
A numeric scalar of price.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
