% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econgoods-package.R
\docType{package}
\name{econgoods-package}
\alias{econgoods}
\alias{econgoods-package}
\title{econgoods: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
