% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goods.R
\name{goods_by}
\alias{goods_by}
\title{Create goods}
\usage{
goods_by(.data, ...)
}
\arguments{
\item{.data}{A data frame that contains columns \code{price} and \code{quantity}.}

\item{...}{Variables to group by.}
}
\value{
A \code{econ_goods} object.
}
\description{
Create goods
}
