% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goods.R
\name{goods_compose}
\alias{goods_compose}
\title{Compose goods}
\usage{
goods_compose(data, utility, node = NULL)
}
\arguments{
\item{data}{A \code{econ_goods} object.}

\item{utility}{A \code{econ_util} object or a data frame that contains columns
\code{utility}.}

\item{node}{A node name for composition goods. By default, it is \code{NULL}.}
}
\value{
A \code{econ_goods} object.
}
\description{
Compose goods
}
