% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goods-reprice.R
\name{goods_reprice}
\alias{goods_reprice}
\title{Reprice goods}
\usage{
goods_reprice(data, prices, gradient = FALSE)
}
\arguments{
\item{data}{A \code{econ_goods} object.}

\item{prices}{A data frame that contains columns \code{price}.}

\item{gradient}{Logical input to return the gradient. By default, \code{FALSE}.}
}
\value{
A \code{econ_goods} object.
}
\description{
Reprice goods
}
