% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util-2goods.R
\name{util_2goods_indifference}
\alias{util_2goods_indifference}
\title{Indifference curve function factory for two goods}
\usage{
util_2goods_indifference(
  f,
  quantities,
  interval = c(1e-06, 1e+06),
  tol = 1e-06,
  ...
)
}
\arguments{
\item{f}{A \code{util} object.}

\item{quantities}{A numeric vector of length 2 with the quantities of the
goods.}

\item{interval}{Passed to \code{\link[stats:uniroot]{stats::uniroot()}}.}

\item{tol}{Passed to \code{\link[stats:uniroot]{stats::uniroot()}}.}

\item{...}{Passed to \code{\link[stats:uniroot]{stats::uniroot()}}.}
}
\value{
A function that takes a scalar numeric of quantity of good X and
returns a scalar numeric of quantity of good Y.
}
\description{
Indifference curve function factory for two goods
}
