% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util_calibrate}
\alias{util_calibrate}
\title{Calibrate a utility function}
\usage{
util_calibrate(f, prices, quantities, ...)
}
\arguments{
\item{f}{A \code{econ_util} object.}

\item{prices}{A numeric vector of prices.}

\item{quantities}{A numeric vector of quantities.}

\item{...}{Additional arguments.}
}
\value{
A \code{econ_util} object with calibrated parameters.
}
\description{
Calibrate a utility function
}
