% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util_expenditure}
\alias{util_expenditure}
\title{Expenditure function}
\usage{
util_expenditure(f, prices, utility, gradient = FALSE, ...)
}
\arguments{
\item{f}{A \code{econ_util} object.}

\item{prices}{A numeric vector of prices.}

\item{utility}{A scalar numeric of utility level.}

\item{gradient}{Logical input to return the gradient. By default, \code{FALSE}.}

\item{...}{Additional arguments.}
}
\value{
When \code{gradient = FALSE}, a scalar numeric of expenditure. When
\code{gradient = TRUE}, a numeric vector of gradients of expenditure related to
prices.
}
\description{
Expenditure function
}
