% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util_indirect}
\alias{util_indirect}
\title{Indirect utility function}
\usage{
util_indirect(f, prices, income, gradient = FALSE, ...)
}
\arguments{
\item{f}{A \code{econ_util} object.}

\item{prices}{A numeric vector of prices.}

\item{income}{A scalar numeric of income.}

\item{gradient}{Logical input to return the gradient. By default, \code{FALSE}.}

\item{...}{Additional arguments.}
}
\value{
When \code{gradient = FALSE}, a scalar numeric of utility level. When
\code{gradient = TRUE}, a numeric vector of gradients of utility level related to
prices.
}
\description{
Indirect utility function
}
